[sg] Simplify generated security-group names
As described in #205 the purpose of this patch is to provide a better out of
the box security group naming convention. The current approach
`[security-group-id]-[security-group-name]` is not very convenient, since we
don't know the security group id in advance. The only reason it's useful is
that it provides 100% unique security group resoruce ids.

A better approach is to use security group names for EC2 Classic, and vpc_id as
prefix for VPC security groups.

This also keeps the unique resource ids but makes them easier to handle
and read.