Adding compatibility with actual terraform version
Yesterday I tried to use terraforming with to get ec2 instances and there was some uncompatibility with new version of terraform. In fact tags aren't declared as resource, but as a map so it changes to "tags = {" instead of "tags {" and keys in this shouldn't be specified as a String. After rebuilding it works correctly